requirements/base.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.6.0
billiard==3.6.4.0
celery==5.0.5
certifi==2021.10.8
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.1.7
djangorestframework==3.12.4
kombu==5.2.1
Pillow==8.1.2
prompt-toolkit==3.0.22
psycopg2-binary==2.8.6
pycodestyle==2.8.0
python-decouple==3.3
pytz==2021.3
sentry-sdk==1.5.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
django-cors-headers==3.5.0
django-filter==2.3.0
djangorestframework-simplejwt==4.5.0
